Renamed calibration types (#595)
