+++ src/csharp/SDK/Extrinsics.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+using System;
+using System.Numerics;
+using System.Runtime.InteropServices;
+
+namespace Microsoft.Azure.Kinect.Sensor
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct Extrinsics
+    {
+        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 9)]
+        public float[] rotation;
+
+        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 3)]
+        public float[] translation;
+    }
+}
